requirements.txt
setuptools
dataclasses
bcrypt
Flask[async]>=2.0.1
Flask-Login>=0.5.0
Flask-SQLAlchemy>=2.5.1
gunicorn[gthread]>=20.1.0
requests>=2.25.1
psutil>=5.8.0